switch to a better method of passing the publish token in client
Rather than storing all the tokens used to publish, we are now wrapping
the data object into another one storing the original data and the token.
That way we don't have to store a list of tokens to use for publishing
based on channel in the client, which was messy and potentially error-prone.