Switch to SQL for the application/key store.
The Dynamo/File abstraction was too unwieldy, and I really wanted
to keep the property that it was easy to run pandapush as a single
dockerfile. Getting rid of that abstraction meant I needed a new one
that would provide both a local datastore that didn't need a server
(sqlite) and a remote one for distributed setups (postgres).

This commit switches the data job to knex, and cleans up a bunch
of crap.